requirements.txt
absl-py==1.4.0
addict==2.4.0
appdirs==1.4.4
asttokens
astunparse==1.6.3
audioread==3.0.0
backcall
backports.functools-lru-cache
basicsr==1.4.2
Bottleneck
brotlipy==0.7.0
cachetools==5.3.0
certifi
cffi
charset-normalizer
clean-fid==0.1.35
click
cryptography
cycler
debugpy
decorator==4.4.2
dlib==19.24.2
docker-pycreds==0.4.0
dominate==2.7.0
einops==0.6.1
entrypoints
executing
face-alignment==1.4.1
facexlib==0.3.0
ffmpeg-python==0.2.0
filelock==3.12.2
filterpy==1.4.5
flatbuffers==23.3.3
flit_core
fonttools==4.25.0
fsspec==2023.6.0
future==0.18.3
fvcore
gast==0.4.0
gfpgan==1.3.8
gitdb==4.0.10
GitPython==3.1.32
glfw==2.2.0
google-auth==2.16.2
google-auth-oauthlib==1.1.0
google-pasta==0.2.0
GPUtil==1.4.0
grpcio==1.51.3
h5py==3.8.0
huggingface-hub==0.16.4
idna
imageio
imageio-ffmpeg==0.4.3
imgcrop==0.1.0
imgui==1.3.0
importlib-metadata==6.8.0
iopath==0.1.9
ipykernel
ipython
jedi
jittor==1.3.6.11
joblib==1.3.1
jupyter-client
jupyter_core
keras==2.11.0
kiwisolver
kornia==0.6.10
lazy_loader==0.1
libclang==15.0.6.1
librosa==0.9.2
llvmlite==0.39.1
lmdb==1.4.1
lpips==0.1.4
Markdown==3.4.1
MarkupSafe==2.1.2
matplotlib
matplotlib-inline
mkl-fft==1.3.1
mkl-random
mkl-service==2.4.0
moviepy==1.0.3
mrcfile==1.4.3
msgpack==1.0.5
mtcnn==0.1.1
munkres==1.1.4
nest-asyncio
networkx==3.0
numba==0.56.4
numexpr
numpy
oauthlib==3.2.2
olefile
opencv-python==4.7.0.72
opt-einsum==3.3.0
packaging
pandas==1.5.2
parso
pathtools==0.1.2
pexpect
pickleshare
Pillow==9.4.0
platformdirs==4.0.0
ply==3.11
pooch==1.6.0
portalocker
positional-encodings==6.0.1
proglog==0.1.10
prompt-toolkit
protobuf==3.19.6
psutil==5.9.4
ptyprocess
pure-eval
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAudio==0.2.14
pycparser
pydantic==1.10.6
Pygments
PyOpenGL==3.1.5
pyOpenSSL
pyparsing
PyQt5-sip==12.11.0
PySocks
pyspng==0.1.1
python-dateutil
python-speech-features==0.6
pytorch-fid==0.3.0
pytorch-msssim==1.0.0
pytz
PyWavelets==1.4.1
PyYAML
pyzmq==19.0.2
realesrgan==0.3.0
regex==2023.8.8
requests
requests-oauthlib==1.3.1
resampy==0.4.2
rsa==4.9
safetensors==0.3.2
scenedetect==0.6.2
scikit-image==0.20.0
scikit-learn==1.3.0
scipy==1.9.1
sentry-sdk==1.29.2
setproctitle==1.3.2
sip
six
smmap==5.0.0
soundfile==0.12.1
soxr==0.3.5
stack-data
tabulate
tb-nightly==2.16.0a20231124
tensorboard==2.11.2
tensorboard-data-server==0.7.2
tensorboard-plugin-wit==1.8.1
tensorflow==2.11.0
tensorflow-estimator==2.11.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor
tf-keras-nightly==2.16.0.dev2023112410
threadpoolctl==3.1.0
tifffile==2023.2.28
tokenizers==0.15.0
toml
tomli==2.0.1
torch==1.13.1+cu116
torchvision==0.14.1+cu116
tornado
tqdm
traitlets
transformers==4.35.2
trimesh==3.20.2
typing_extensions
urllib3
wandb==0.15.8
wcwidth
Werkzeug==2.2.3
wrapt==1.15.0
yacs
yapf==0.40.2
zipp==3.15.0

# simple - Tested on RTX 3090, ubuntu 18.04
# conda create -n talk3d python=3.8
# conda activate talk3d
# pip install torch==1.12.1+cu116 torchvision==0.13.1+cu116 torchaudio==0.12.1 --extra-index-url https://download.pytorch.org/whl/cu116
# pip install absl-py basicsr cachetools clean-fid click cryptography dlib einops facexlib ffmpeg-python fvcore gfpgan h5py imageio imageio-ffmpeg ipykernel joblib keras librosa lmdb lpips matplotlib moviepy mrcfile mtcnn opencv-python pandas Pillow ply pydantic pytorch-fid pytorch-msssim realesrgan requests scipy scikit-learn six soundfile tensorboard tensorflow tqdm wandb yacs trimesh transformers kornia positional-encodings[pytorch] face_alignment